dom/indexedDB/test/test_filehandle_ordering.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
  function* testSteps()
  {
    const name = window.location.pathname;

    let request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    let db = event.target.result;
    db.onerror = errorHandler;

    request = db.createMutableFile("test.txt");
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let mutableFile = event.target.result;
    mutableFile.onerror = errorHandler;

    let fileHandle1 = mutableFile.open("readwrite");
    let fileHandle2 = mutableFile.open("readwrite");

    fileHandle2.write("2");
    fileHandle1.write("1");

    fileHandle1.oncomplete = grabEventAndContinueHandler;
    fileHandle2.oncomplete = grabEventAndContinueHandler;

    yield undefined;
    yield undefined;

    let fileHandle3 = mutableFile.open("readonly");
    let request3 = fileHandle3.readAsText(1);
    request3.onsuccess = grabEventAndContinueHandler;

    event = yield undefined;
    is(event.target.result, "2", "File handles were ordered properly.");

    finishTest();
  }
  </script>
  <script type="text/javascript" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
